pyproject.toml
[project]
name = "discord-event"
version = "0.1.0"
description = "The Iron Foundry — Event management Discord bot"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "discord-py>=2.7.1",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "pydantic>=2.12.3",
    "pymongo>=4.15.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
